requirements.txt
python-nmap==0.6.1
python-netdiscover==0.2.0
mac-vendor-lookup==0.1.8